% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_csv.R
\name{export_ufc_data_csv}
\alias{export_ufc_data_csv}
\title{Export UFC data for a fighter to a CSV}
\usage{
export_ufc_data_csv(fighter_name, file_path)
}
\arguments{
\item{fighter_name}{Character. Name of the fighter.}

\item{file_path}{Character. Path to the CSV file.}
}
\description{
Export UFC data for a fighter to a CSV
}
\examples{
\dontrun{
  # This example fetches data for Israel Adesanya and saves it to a CSV file
  export_ufc_data_csv("Israel Adesanya", "Israel_Adesanya_Fight_Data.csv")
}

}
