% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper_functions.R
\name{get_ufc_data}
\alias{get_ufc_data}
\title{Get UFC Fight Data for a Given Fighter}
\usage{
get_ufc_data(fighter_name)
}
\arguments{
\item{fighter_name}{Character. Name of the fighter, e.g. "Israel Adesanya".}
}
\value{
A data frame of fight data for the specified fighter.
}
\description{
Get UFC Fight Data for a Given Fighter
}
\examples{
\dontrun{
  # This example fetches data for Israel Adesanya.
  df <- get_ufc_data("Israel Adesanya")
  head(df)
}

}
